More specific with button placement
If a mod adds a new section to the top of the options menu, and that mod is loaded before this one, the reload buttons would try to be added to the other mod's section. This update fixes that behavior by targeting the section with the Change Language button, which is the intended one.